store maxSendQBytes in a uint32 to avoid alignment problems
The sync.atomic documentation says:

"On both ARM and x86-32, it is the caller's responsibility to arrange
for 64-bit alignment of 64-bit words accessed atomically."

+++ irc/server.go
@@ -109,7 +109,7 @@ type Server struct {
 	limits                     Limits
 	listeners                  map[string]*ListenerWrapper
 	logger                     *logger.Manager
-	maxSendQBytes              uint64
+	maxSendQBytes              uint32
 	monitorManager             *MonitorManager
 	motdLines                  []string
 	name                       string
